+++ Inc/kobjs.h
@@ -112,7 +112,7 @@
  {
      RK_KOBJ_ID objID;
      BOOL init;
-     LONG value;
+     INT value;
      struct kTcb *owner;
      struct kList waitingQueue;
  } __attribute__((aligned(4)));
